Move CCDF variables into their own files
