+++ openfisca_us/variables/hhs/ccdf/is_ccdf_initial_income_eligible.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class is_ccdf_initial_income_eligible(Variable):
+    value_type = bool
+    entity = SPMUnit
+    definition_period = YEAR
+    label = u"Initial income eligibility for CCDF"
