+++ openfisca_us/variables/hhs/ccdf/is_enrolled_in_ccdf.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class is_enrolled_in_ccdf(Variable):
+    value_type = bool
+    entity = Person
+    definition_period = YEAR
+    label = u"CCDF enrollment status"
