+++ openfisca_us/variables/hhs/ccdf/spm_unit_total_childcare_market_rate.py
@@ -1,0 +1,12 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_total_childcare_market_rate(Variable):
+    value_type = float
+    entity = SPMUnit
+    definition_period = YEAR
+    label = u"SPM unit total childcare market rate"
+    unit = "currency-USD"
+
+    def formula(spm_unit, period, parameters):
+        return spm_unit.sum(spm_unit.members("ccdf_market_rate", period))
